File->Open pops up a modal 'open files' dialog
The File->Open dialog lets the user choose one or more files which are
opened in tabs in a new window; it replaces the behaviour where
File->Open presented a sheet which opened (at most) one file in the
topmost window.  This conforms better with the standard File->Open
behaviour of Mac apps.