File opening procedure changed
When files is being opened (via [MMAppController
application:openFiles:]) try to open the files in an untitled window
(either one that is already open, or one that is currently launching).
This avoids two windows opening when the 'Edit in MacVim' input manager
is used, or when files are opened from Xcode.