Preference panel update, input manager installer
The preference panel now uses Dave Batton's DBPrefsWindowController
class.  A (preliminary) "Integration" panel has also been added which
adds buttons to install the "Edit in ODBEditor" input manager, and a way
to choose which editor to use with the input manager.

The input manager is now bundled inside MacVim.app and the source code
has been added to the project.  The original ODB input manager source
code was written by Allan Odgaard and was later modified by Chris Eidhof
and Eelco Lempsink to make it work with other editors than TextMate.

The installer uses a new AuthorizedShellCommand class.

Dave Batton is credited in the About dialog.