Improved method to start Vim processes in a login shell
A login shell is started by exec'ing a shell process whose argv[0] is
prepended with a dash.  Which shell to use can be controlled with user
default MMLoginShellCommand (or by setting $SHELL).  If the shell needs
a parameter to make it a login shell then this can be set with the user
default MMLoginShellArgument.  The documentation has been updated with
these new user defaults.