vim_to_gauche(): Don't use Scm_Error directly
Scm_Error and others do jump to an error handler, so the rest code of a caller
of Scm_Error will not be executed.  It causes some leaks of memory.  Fix
vim_to_gauche() to return an error message and not to use Scm_Error and others
directly.

Note that gauche_to_vim(), vim_apply_proc() and vim_eval_proc() have the same
problem.  They have to be fixed later.